# api/api.yaml
info:
  title: Veteran Support API
  version: 1.0.0
host: 'localhost:5000'
basePath: /
swagger: '2.0'
paths:
  /resources:
    post:
      description: create new resource
      parameters:
        - in: body
          name: body
          description: Resource Object
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              ok:
                type: boolean
              id:
                type: string
              rev:
                type: string
        '400':
          description: Missing body or required fields
          schema:
            type: object
            properties:
              ok:
                type: boolean
              message:
                type: string
    get:
      description: list resources
      parameters:
        - in: query
          name: filter
          type: string
          required: false
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Resource'
  '/resources/{id}':
    get:
      description: get resource by id
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Resource'
        '500':
          description: Error
    put:
      description: update resource
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: body
          name: body
          description: Resource Object
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              ok:
                type: boolean
              id:
                type: string
              rev:
                type: string
        '400':
          description: Missing body or required fields
          schema:
            type: object
            properties:
              ok:
                type: boolean
              message:
                type: string
    delete:
      description: remove resource by id
      parameters:
        - in: path
          name: id
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              ok:
                type: boolean
  /categories:
    post:
      description: >-
        Adds a category to the collection of categories. When adding a category
        the `name`, `short desc`, and `desc` are required fields.
      parameters:
        - in: body
          name: body
          description: Category Object
          required: true
          schema:
            $ref: '#/definitions/Category'
      responses:
        '200':
          description: Success
          schema:
            type: object
            properties:
              ok:
                type: boolean
              id:
                type: string
              rev:
                type: string
        '400':
          description: Missing body or required fields
          schema:
            type: object
            properties:
              ok:
                type: boolean
              message:
                type: string
definitions:
  Resource:
    type: object
    properties:
      _id:
        type: string
      categoryId:
        type: string
        description: reference identifier to category doc
      name:
        type: string
        description: name of resource for veteran
      formalName:
        type: string
        description: Long Name description of resource
      shortDesc:
        type: string
        description: Short Name of resource
      purpose:
        type: string
        description: long description of resource
      website:
        type: 'string:'
        description: url of website
      primaryPhone:
        type: string
        description: primary contact phone number
      rank:
        type: integer
        description: resource rank
      faq:
        type: array
        items:
          $ref: '#/definitions/Faq'
  Faq:
    type: object
    properties:
      _id:
        type: string
      question:
        type: 'string:'
        description: Frequently Asked Question
      answer:
        type: string
        description: Answer to question
  Category:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
        description: name of category
      shortDesc:
        type: string
        description: short description of category
      icon:
        type: string
        description: icon class name
responses: {}
parameters: {}
securityDefinitions: {}
tags: []